pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ms-ocr"
version = "0.1.0"
description = "High-quality PDF text extraction with OCR, focused on Spanish content"
authors = [{name = "Steven Ayal", email = "steven@aiquaa.com"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["ocr", "pdf", "text-extraction", "spanish", "gamma", "markdown"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Text Processing",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "PyMuPDF>=1.23.8",
    "pytesseract>=0.3.10",
    "opencv-python>=4.9.0",
    "Pillow>=10.2.0",
    "layoutparser>=0.3.4",
    "camelot-py[cv]>=0.11.0",
    "tabula-py>=2.9.0",
    "unidecode>=1.3.7",
    "langdetect>=1.0.9",
    "python-docx>=1.1.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "tqdm>=4.66.1",
    "structlog>=24.1.0",
    "pydantic>=2.5.3",
    "jsonschema>=4.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.14",
    "black>=23.12.1",
    "isort>=5.13.2",
    "mypy>=1.8.0",
]

[project.scripts]
ms-ocr = "ms_ocr.cli:main"

[project.urls]
Homepage = "https://github.com/stevenayal/ms_ocr"
Repository = "https://github.com/stevenayal/ms_ocr"
Issues = "https://github.com/stevenayal/ms_ocr/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/ms_ocr --cov-report=term-missing"
